[REFACTOR]: extract library state into a custom hook
